feat: 新增url_launcher依赖，更新版本并优化歌词功能
1. 新增url_launcher相关依赖及Windows平台插件支持
2. 升级应用版本到2.0.0+2
3. 重构歌词获取逻辑，新增lrc格式歌词获取
4. 优化安卓桌面歌词通知栏图标和悬浮歌词布局
5. 移除歌词显示开关设置，新增版本更新跳转链接

+++ android/app/src/main/kotlin/com/md3music/md3music/MainActivity.kt
@@ -99,8 +99,8 @@ class MainActivity : FlutterActivity() {
                 "setDesktopLyricConfig" -> {
                     val intent = Intent(this, FloatingLyricService::class.java).apply {
                         action = FloatingLyricService.ACTION_SET_CONFIG
-                        call.argument<Float>(FloatingLyricService.EXTRA_FONT_SIZE)?.let {
-                            putExtra(FloatingLyricService.EXTRA_FONT_SIZE, it)
+                        call.argument<Double>(FloatingLyricService.EXTRA_FONT_SIZE)?.let {
+                            putExtra(FloatingLyricService.EXTRA_FONT_SIZE, it.toFloat())
                         }
                         call.argument<Boolean>(FloatingLyricService.EXTRA_DOUBLE_LINE)?.let {
                             putExtra(FloatingLyricService.EXTRA_DOUBLE_LINE, it)
